requirements.txt
fastapi>=0.103.2
uvicorn>=0.23.2
python-multipart
pytesseract
Pillow
pydantic>=2.5.0
pytest
httpx
slowapi
